requirements.txt
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.110.0
pre-commit==3.6.2
pydantic==2.5.2
pydantic-settings==2.0.3
pydantic_core==2.14.5
python-decouple==3.8
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
uvicorn==0.27.1
gunicorn